+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   - [#1519](https://github.com/wix-incubator/rich-content/pull/1519) external toolbars: inline style toggling fixed
 - `plugin-link`
   - [#1521](https://github.com/wix-incubator/rich-content/pull/1521) fix basic link panel wiring
+- `ricos-viewer`
+  - [#1522](https://github.com/wix-incubator/rich-content/pull/1522) incorrect gallery image index in fullscreen
 
 ### :home: Internal
 - `e2e`